# path: refactor for performance and consistency