Improve performance by:
+ Not leaking the `arguments` object! (in [`win32.join`](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881L197))
+ Getting the last character of a string by index, instead of with `.substr()` ([here](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881L445)) or `.slice()` ([here](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881L363))

Improve code consistency by:
+ Using `[]` instead of `.charAt()` where possible (such as [here](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881L428))
+ Using a function declaration instead of a var declaration ([here](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881R114))
+ Using `.slice()` with clearer arguments (such as [here](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881L387))

Improve both by:
+ Making the reusable `trimArray()` function
+ Standardizing getting certain path statistics with the new `win32StatPath()` function

#### Benchmarks: (higher is better)
| function (path.) | current (ops/sec) | this PR (ops/sec)
| -----------------|-------------------|------------------
| win32.resolve | 315,019 | 399,543
| win32.normalize | 873,559 | 1,051,170
| win32.isAbsolute | 1,915,332 | 1,939,849
| win32.join | 452,879 | 815,033
| win32.relative | 153,821 | 219,928
| win32.format | 13,118,588 | 16,364,384
| posix.relative | 244,638 | 304,456
##### [Benchmark code (gist)](https://gist.github.com/woollybogger/6535c360ca196fe0e51e)