@woollybogger 2015-03-04
@jasnell I updated the PR description with benchmark data. I only included the functions that had a >1% change in performance.