@jasnell 2015-03-12
@woollybogger ... sorry, was catching up on a few other items. Thanks for the profiling numbers. I'll queue this up to review on Monday. 

// cc @joyent/node-coreteam ... can one of you take a look as well?