@piscisaureus 2015-03-16
I know the damage has already been done, but one of the objections I have against exposing the unix code on windows and vice versa is that `path.resolve()` and maybe other functions produce garbage when relative paths are used.

Does anyone have an idea for fixing this. Maybe throw if the path is relative?