@misterdjules 2015-03-17
@piscisaureus I'm not sure I understand what you're saying, could you please give us an example of what you're describing?