@piscisaureus 2015-03-17
@misterdjules 

```
> path.posix.resolve('foo/bar')
'C:\\Users\\Bert Belder/foo/bar'
```

I don't know what would happen if you did the opposite on unix, e.g.
```
> path.win32.resolve('\\\\myserver\\myshare\\foo.txt');
???
> path.win32.resolve('c:appdata\\node.js');
???
```

But I am sure the results would not make sense.