@misterdjules 2015-03-17
@piscisaureus OK thank you for the clarification!

```
> path.win32.resolve('\\\\myserver\\myshare\\foo.txt');
???
> path.win32.resolve('c:appdata\\node.js');
???
```
work as expected because they are absolute paths, but like you said `path.win32.resolve` with relative paths would fail on any UNIX system.

It seems that only `path.resolve` would be affected by this, as it seems to be the only one calling a platform-specific function (`process.cwd`). `path.relative` uses `path.resolve`, but it doesn't seem to be affected by it since it seems to deal only with relative paths. We would need to investigate more though.

Anyway, I think this should be handled outside of this PR and in my opinion it's worth creating a new issue. Do you want to do that @piscisaureus or should I do it?

Thank you again!