@piscisaureus 2015-03-17
@misterdjules 

Ah, oops, you're right.
(The second path `c:appdata\\node.js` isn't absolute btw). It does a drive-specific cwd lookup [here](https://github.com/joyent/node/blob/9b534e2e87489d219f0d0cf5be7f9119704fe424/lib/path.js#L102).

Agreed btw that I should be handled in a separate issue.