@woollybogger 2015-03-19
@misterdjules @jasnell Updated with the helper function as discussed [here](https://github.com/joyent/node/pull/9289#discussion-diff-26521014) and updated the benchmark data.