@misterdjules 2015-03-19
Could you please take a look at my [comment](https://github.com/joyent/node/pull/9289#discussion_r26788172)? Otherwise, LGTM. Thanks @woollybogger! :+1:

Adding to the 0.12.2 milestone as node v0.12.1 will be a security release following the recent OpenSSL releases.