@woollybogger 2015-03-20
@roryrjb You wrote the `path.format` and `path.parse` functions so I have a question for you. In `path.format`, [`root`](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881R371) isn't used anywhere except to [`throw an error`](https://github.com/joyent/node/pull/9289/files#diff-c0fae59301ef49fb71b43278db9ad881R373). Is there a good reason for this or can that check just be removed?