@roryrjb 2015-03-20
@woollybogger no, there's no reason I can think of, it can be removed.