@misterdjules 2015-03-23
@woollybogger If we want to make changes to `path.format`, let's do that on master and in another issue/PR. The documentation in v0.12 doesn't mention a specific behavior for any platform, and it's not clear how many users currently rely on that behavior.

Thank you very much again!