@woollybogger 2015-03-24
@misterdjules Added a new commit with the tests and a couple other changes related to `path.parse` and `path.format`. If it's good I'll squash the commits.