@misterdjules 2015-03-24
Thank you again @woollybogger. Added some comments, please let me know what you think. 

For some of the more controversial changes, we can always tackle them later in separate PRs. That way we can land the improvements you already made to the path's module performance and take more time to discuss the rest.