@woollybogger 2015-03-25
@misterdjules I removed the controversial code and squashed into one commit.